changed directory names and corresponding imports
